# Async callback memory leak fix 2