… akka-bench-jmh/src/main/scala/akka/stream/InvokeWithFeedbackBenchmark.scala
… akka-stream-tests/src/test/scala/akka/stream/impl/fusing/AsyncCallbackSpec.scala
… akka-stream-tests/src/test/scala/akka/stream/scaladsl/SubStreamMemoryLeaks.scala