Based on Johannes idea to keep track of async callbacks based on `invokeWithFeedback` usage.

Fixes #24046