@johanandren 2017-11-29
Benchmarked to see that it's not a crazy cost to read that volatile on each invoke:

```
With auto-cleaning:
[info] Benchmark                                           (NumberOfStreams)   Mode  Cnt      Score     Error   Units
[info] FlatMapMergeBenchmark.flat_map_merge_100k_elements                  0  thrpt  200  20707.947 ± 163.854  ops/ms
[info] FlatMapMergeBenchmark.flat_map_merge_100k_elements                  1  thrpt  200   4796.247 ±  34.859  ops/ms
[info] FlatMapMergeBenchmark.flat_map_merge_100k_elements                 10  thrpt  200   4555.953 ±  41.906  ops/ms

master:
[info] Benchmark                                           (NumberOfStreams)   Mode  Cnt      Score     Error   Units
[info] FlatMapMergeBenchmark.flat_map_merge_100k_elements                  0  thrpt  200  20767.491 ± 209.263  ops/ms
[info] FlatMapMergeBenchmark.flat_map_merge_100k_elements                  1  thrpt  200   4749.419 ±  21.732  ops/ms
[info] FlatMapMergeBenchmark.flat_map_merge_100k_elements                 10  thrpt  200   4716.884 ±  41.707  ops/ms
```