@johanandren 2017-11-29
Reproducer

before (master):
<img width="539" alt="screen shot 2017-11-28 at 14 52 21" src="https://user-images.githubusercontent.com/666915/33383306-3fc43742-d523-11e7-93d8-6ff2387f17a7.png">

after:
<img width="534" alt="screen shot 2017-11-29 at 16 34 31" src="https://user-images.githubusercontent.com/666915/33383305-3faab132-d523-11e7-90ae-feeea6f3d185.png">