@johanandren 2017-11-30
Benchmark for the latest commit also looks on par with master.