@johanandren 2017-12-05
Verified again that there is no memory leak anymore. I also ran benchmarks yesterday, it looks a tiny bit worse than master so I'd like to profile and see if there is a cost I didn't think of:

this branch with the extra atomic-ref-touch removed (at 033b936)
[info] Benchmark                                           (NumberOfStreams)   Mode  Cnt      Score     Error   Units
[info] FlatMapMergeBenchmark.flat_map_merge_100k_elements                  0  thrpt  200  19858.650 ± 209.228  ops/ms
[info] FlatMapMergeBenchmark.flat_map_merge_100k_elements                  1  thrpt  200   4739.656 ±  35.219  ops/ms
[info] FlatMapMergeBenchmark.flat_map_merge_100k_elements                 10  thrpt  200   3994.138 ± 112.990  ops/ms


master:
[info] Benchmark                                           (NumberOfStreams)   Mode  Cnt      Score     Error   Units
[info] FlatMapMergeBenchmark.flat_map_merge_100k_elements                  0  thrpt  200  20462.595 ± 191.099  ops/ms
[info] FlatMapMergeBenchmark.flat_map_merge_100k_elements                  1  thrpt  200   4673.038 ±  27.722  ops/ms
[info] FlatMapMergeBenchmark.flat_map_merge_100k_elements                 10  thrpt  200   4706.538 ±  55.010  ops/ms