@johanandren 2017-12-05
Latest in this branch and master compared through the added invoke-with-feedback-bench, I'm happy with this compared to the previous one.

```
branch (939a6e3)
[info] Benchmark                                                Mode  Cnt       Score      Error  Units
[info] InvokeWithFeedbackBenchmark.pass_through_100k_elements  thrpt   40  115312.862 ± 5579.883  ops/s

master
[info] Benchmark                                                Mode  Cnt       Score      Error  Units
[info] InvokeWithFeedbackBenchmark.pass_through_100k_elements  thrpt   40  112832.329 ± 1518.963  ops/s
```