@patriknw 2017-12-05
There is another benchmark testing the `Source.queue` and I see a regression compared to v2.5.6, but perhaps that is the price we have to pay for feedback and Source.queue is known to not be very high performance.

jmh:run -prof gc -i 7 -wi 3 -f1 -t1 akka.remote.artery.SendQueueBenchmark

```
this branch:
[info] SendQueueBenchmark.queue                                      thrpt    7   697.773 ±  113.748  ops/ms
[info] SendQueueBenchmark.queue:·gc.alloc.rate                       thrpt    7   621.601 ±  270.570  MB/sec

v2.5.7:
[info] SendQueueBenchmark.queue                                      thrpt    7   930.796 ±  218.450  ops/ms
[info] SendQueueBenchmark.queue:·gc.alloc.rate                       thrpt    7   375.063 ±  135.179  MB/sec

v2.5.6:
[info] SendQueueBenchmark.queue                                       thrpt    7   2041.608 ± 394.991  ops/ms
[info] SendQueueBenchmark.queue:·gc.alloc.rate                        thrpt    7    228.065 ±  45.148  MB/sec
```

No regression in the Artery SendQueue, which is using ordinary `invoke` (sparingly).