@johanandren 2017-12-05
Damned sbt NPE...