@johanandren 2017-12-05
Tried avoiding that `OptionVal` boxing (because of passing `OptionVal` to a lambda in `onAsyncInput`, strangely enough that made it a bit slower, tried to avoid an extra allocation for every update to the map because of `Map$WithDefault` - did also not help any with throughput. I think I'm done trying to optimize this for now.