@jrudolph 2017-12-05
I think the main reason is probably that there's lots of contention on those shared data structures so that the slow map/set operations are probably executed multiple times.

I tried to use `synchronized` while accessing `asyncCallbacksInProgress` instead of the `AtomicReference` which isn't slower than what we have now which seems like a confirmation of the above suspicion.

So, I guess the only way to better performance would be to remove contention. Two kinds of contention will happen:
 * two promises being added at the same time (not tested in the benchmark iiuc)
 * a promise being added and one being removed

We cannot do much about the first one aside from using a good concurrent data structure to store the entries. To improve the second situation we would probably have to amortize the removal operation by running it regularly (I think you suggested that at one point, @johanandren).